Implement collision handling.
When bodes are found to be overlapping, compute whether they are
still moving towards each other.  If so, absorb the normal velocity
component and preserve the portion of each body's velocity along the
tangent between the bodies.  Stop applying gravity until bodies are
no longer overlapping.

The main problem seems to be that we don't catch a collision until
after the bodies have started overlapping.  Pausing gravity until
they're done separating helps compensate for that.

One advantage of this approach is that we don't have to remember
which bodies are in the process of colliding.